requirements.txt
beautifulsoup4==4.12.2
numpy==1.26.3
pandas==2.1.4
price-parser==0.3.4
pymongo==4.6.1
ruamel.yaml==0.18.5
selenium==4.16.0
requests==2.31.0
webdriver_manager-4.0.1
free-proxy-1.1.1